+++ src/features/walletConnect/menu/Connected.tsx
@@ -1,4 +1,4 @@
-import MenuProfile from '@features/profile/menu';
+import MenuProfile from '@features/profile/menu/layout';
 
 export default function MenuConnected() {
   return (
