+++ src/recoil/connected/address/atoms.ts
@@ -10,6 +10,10 @@ export const addressConnectedAtom = atom({
   default: ''
 });
 
+export const balanceBankConnectedAtom = atom({
+  key: 'balanceBankConnectedAtom'
+});
+
 export const balanceYfdConnectedAtom = atom({
   key: 'balanceYfdConnectedAtom',
   default: '0'
@@ -19,9 +23,3 @@ export const balanceFyfdConnectedAtom = atom({
   key: 'balancefYfdConnectedAtom',
   default: '0'
 });
-
-export default {
-  addressConnectedAtom,
-  balanceYfdConnectedAtom,
-  balanceFyfdConnectedAtom
-};
